shared test harness / worker
