+++ packages/cosmos/test/client.integration.spec.ts
@@ -1,5 +1,4 @@
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
 import {
   CosmosClient,
   IndexingPolicy,
