+++ packages/cosmos/test/index.ts
@@ -1,0 +1,7 @@
+import worker from '../../../test-worker.js';
+import './client.integration.spec.js';
+import './retry.spec.js';
+import './session.spec.js';
+import './signer.spec.js';
+import './util.spec.js';
+export default worker;
