+++ packages/cosmos/test/util.spec.ts
@@ -1,5 +1,4 @@
 import { expect } from 'chai';
-import { describe, it } from 'mocha';
 import {
   assertArg,
   escapeNonASCII,
