+++ packages/web/test/index.ts
@@ -1,4 +1,4 @@
-import './_mocha-setup.js';
+import worker from '../../../test-worker.js';
 import './accepts.spec.js';
 import './application.spec.js';
 import './context.spec.js';
@@ -7,20 +7,4 @@ import './normalize-pathname.spec.js';
 import './req.spec.js';
 import './response-builder.spec.js';
 import './validate.spec.js';
-
-let running = false;
-
-export default {
-  async fetch() {
-    if (running) {
-      return new Response('Already running', { status: 400 });
-    }
-    running = true;
-
-    const result = await new Promise(resolve => mocha.run(resolve));
-    return new Response(JSON.stringify(result, null, 2), {
-      status: 200,
-      headers: { 'content-type': 'application/json' }
-    });
-  }
-};
+export default worker;
