+++ test-worker.ts
@@ -18,3 +18,20 @@ globalThis.location = {
 
 // @ts-ignore
 mocha.checkLeaks();
+
+let running = false;
+
+export default {
+  async fetch() {
+    if (running) {
+      return new Response('Already running', { status: 400 });
+    }
+    running = true;
+
+    const result = await new Promise(resolve => mocha.run(resolve));
+    return new Response(JSON.stringify(result, null, 2), {
+      status: 200,
+      headers: { 'content-type': 'application/json' }
+    });
+  }
+};
