+++ test.mjs
@@ -1,7 +1,7 @@
 import { unstable_dev } from 'wrangler';
 
 try {
-  const worker = await unstable_dev('dist/index.js', {
+  const worker = await unstable_dev(process.cwd() + '/dist-test/index.js', {
     local: true,
     logLevel: 'info',
     experimental: { disableExperimentalWarning: true }
